+++ snax/data_sources/in_memory_data_source.py
@@ -2,7 +2,7 @@ from typing import Optional, Dict, List
 
 import pandas as pd
 
-from snax.data_sources.data_source import DataSourceBase
+from snax.data_sources.data_source_base import DataSourceBase
 
 
 class InMemoryDataSource(DataSourceBase):
